requirements.txt
Flask==2.0.1
firebase_admin==5.1.0
python-dotenv==0.19.1
aiohttp==3.8.1
requests==2.26.0
pandas==1.4.0
flask-cors==3.0.10
werkzeug==2.3.0
pydantic>=1.9.0
urllib3>=2.0.7

